Okay, I think my python bindings have reached the border of useful, so in the svn they go.
The bindings currently allow for adb create and open, status and the setting of the l2norm and power flags.

Note that these are the direct c bindings, when it's done there will be OO dressing on top for standard entry.
Also, even though I don't use numpy yet, the include file is brought in, so the module won't build if you don't have numpy in an importable place.



git-svn-id: svn+ssh://rvw/isms/group/omras2/src/svn/audioDB/trunk@930 7f337319-b335-0410-8187-a933a6394831